Feature flags for pools and markets (#1140)
* feature flags

* add events

* core module tests

* legacy market cannonfile

* test fixes

* fix bytes32 spotmarket

* fix naming

* add feature flags in v3 by default

* wip refactor

* refactor to use libraries

* add feature flag to cannonfile

* remove permission load

* remove router

* change pool to market

* fix cannon

* add multiple addresses feature flag

* lint

* change overloaded functions

* code review fixes

* fix feature flag module addition to routers

* remove feature flag from some systems

* move feature flag

* noah changes

* add utility funcs

* lint and rename

* remove comments from config

+++ markets/spot-market/contracts/modules/SpotMarketModule.sol
@@ -76,7 +76,7 @@ contract SpotMarketModule is
         return _getTotalSupply().mulDecimal(_getCurrentPrice());
     }
 
-    function name(uint marketId) external view returns (string memory) {
+    function name(uint128 marketId) external view returns (string memory) {
         return string(string.concat(bytes(_getName()), " Spot Market"));
     }
